requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1 # Driver para MySQL
python-dotenv==1.0.1 # Opcional, para gerenciar variáveis de ambiente
Flask-Login==0.6.3 # Para gerenciar autenticação de usuários
